+++ src/apphosting/backend.spec.ts
@@ -106,8 +106,8 @@ describe("apphosting setup functions", () => {
         projectId,
         location,
         backendId,
-        cloudBuildConnRepo,
         "custom-service-account",
+        cloudBuildConnRepo,
         webAppId,
       );