+++ src/deploy/functions/release/index.ts
@@ -196,7 +196,8 @@ async function setupArtifactCleanupPolicies(
         `${locationsWithErrors.join(", ")}.` +
         "This could result in a small monthly bill as container images accumulate over time.",
     );
-    throw new FirebaseError(
+    utils.logLabeledWarning(
+      "functions",
       `Functions successfully deployed but could not set up cleanup policy in ` +
         `${locationsWithErrors.length > 1 ? "regions" : "region"} ${locationsWithErrors.join(", ")}. ` +
         `Pass the --force option to automatically set up a cleanup policy or ` +